+++ CMakeLists.txt
@@ -12,10 +12,10 @@ add_executable(${PROJECT_NAME}
 target_compile_features(
     kidle PRIVATE cxx_std_20)
 
-add_subdirectory(models)
+add_subdirectory(model)
 
 target_link_libraries(${PROJECT_NAME}
                       ${CURSES_LIBRARIES}
                       pthread
-                      models
+                      model
                       )
