+++ model/generator.cpp
@@ -1,4 +1,4 @@
-#include <models/generator.h>
+#include <model/generator.h>
 
 namespace model
 {
